projects/Autogen-Agent-Chat-Planner/pyproject.toml
[project]
name = "autogen-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "streamlit>=1.30,<2.0",
  "autogen-agentchat>=0.2,<0.3",
  "docker>=6,<7",
  "requests>=2.31,<3",
  "python-dotenv>=1.0,<2",
]
